Renamed Utils and PrefabDependencyTreeGenerator more appropriately
